% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_mm.R
\name{power_mm}
\alias{power_mm}
\title{power_mm}
\usage{
power_mm(
  Formula,
  Varcomp,
  Data,
  Effect,
  Family = "Gaussian",
  Alpha = 0.05,
  Resid_var = NULL,
  tol = 1e-06
)
}
\arguments{
\item{Formula}{The formula to be passed into blmer.}

\item{Varcomp}{A vector of variance components to be fixed into the random
effects model. Values will fixed based on the order of random effects
specified in the formula statement.}

\item{Data}{The dataset to be used in blmer.}

\item{Effect}{Fixed effects of interest for the power calculation. Can be a
single value or a vector containing fixed effects specified in the 'Formula'
statement.}

\item{Family}{Specifies the distribution of the data. The default is "Gaussian".}

\item{Alpha}{The critical value for the F-statistic. The default is 0.05.}

\item{Resid_var}{The residual variance value to be fixed. If the example does
not have a residual variance, leave the default as NULL.}

\item{tol}{The tolerance or standard deviation to be used for the covariance
priors. The default is 1e-6.}
}
\description{
Calculates the power of a mixed effects model.
}
